remove codegen options
also split out the AutoPlugin into a separate file
and remove the `TwinaglePlugin` plugin; I'm not sure
why it existed in the first place.
